Add DemoController and /demo route
This route will stream a simple JSON response over 10 seconds. The
`render_stream` helper attaches a Ruby Enumerator to the response body.
When the body is then constructed by the server, it iterates over the
Enumerator resulting in a chunked response. The streaming behavior can
be seen in a browser as the JSON response slowly comes in and the
browser displays the new data as it is received.